refactor: introduce Connection interface
